+++ src/ws_server/index.ts
@@ -1,8 +1,9 @@
 import { RawData, WebSocketServer } from 'ws'
 import { db } from './db'
-import { RESPONSE_TYPES, ERROR_MESSAGES } from './constants'
-import { toResponse, parseRawData } from './utils'
-import { RegResponseI, CreateGameResponseI, UpdateRoomResponseI } from './types'
+import { RESPONSE_TYPES } from './constants'
+import { parseRawData } from './utils'
+import { updateRoom, createRoom, addUserToRoom } from './messages/rooms'
+import { userNotFound, invalidPassword, successLogin } from './messages/users'
 
 export const createWsServer = (port: number): void => {
   const wss = new WebSocketServer({ port })
@@ -20,112 +21,27 @@ export const createWsServer = (port: number): void => {
           const user = db.users.get(ws)
 
           if (!user) {
-            const userToAdd = {
-              index: db.users.size,
-              name: body.name,
-              password: body.password,
-            }
-            db.users.set(ws, userToAdd)
-            const data: RegResponseI = {
-              name: userToAdd.name,
-              index: userToAdd.index,
-              error: false,
-              errorText: '',
-            }
-            const response = toResponse<RegResponseI>(RESPONSE_TYPES.REG, data)
-            ws.send(response)
+            userNotFound(ws, body, db)
             break
           }
 
           if (user && user.password !== body.password) {
-            const data: RegResponseI = {
-              name: user.name,
-              index: user.index,
-              error: true,
-              errorText: ERROR_MESSAGES.INVALID_PASSWORD,
-            }
-            const response = toResponse<RegResponseI>(RESPONSE_TYPES.REG, data)
-            ws.send(response)
+            invalidPassword(ws, user)
             break
           }
 
-          const data: RegResponseI = {
-            name: user.name,
-            index: user.index,
-            error: false,
-            errorText: '',
-          }
-          const response = toResponse<RegResponseI>(RESPONSE_TYPES.REG, data)
-          ws.send(response)
-
-          const rooms = Object.values(db.rooms).map(r => ({ roomId: r.id, roomUsers: r.users }))
-          const updateRoomResponse = toResponse<UpdateRoomResponseI>(
-            RESPONSE_TYPES.UPDATE_ROOM, 
-            rooms
-          )
-          ws.send(updateRoomResponse)
-
+          successLogin(ws, user)
+          updateRoom(ws, db)
           break
         }
         case RESPONSE_TYPES.CREATE_ROOM: {
-          const roomId = Object.keys(db.rooms).length
-          const user = db.users.get(ws)
-          if (user) {
-            db.rooms[roomId] = {
-              id: roomId,
-              users: [{ index: user.index, name: user.name }],
-            }
-          }
-
-          const rooms = Object.values(db.rooms).map(r => ({ roomId: r.id, roomUsers: r.users }))
-          const updateRoomResponse = toResponse<UpdateRoomResponseI>(
-            RESPONSE_TYPES.UPDATE_ROOM, 
-            rooms
-          )
-          ws.send(updateRoomResponse)
+          createRoom(ws, db);
+          updateRoom(ws, db);
           break;
         }
         case RESPONSE_TYPES.ADD_USER_TO_ROOM: {
-          const roomId = body.indexRoom
-          const room = db.rooms[roomId]
-          const [{ index: firstUserId }] = room.users
-          const firstUserWs = db.findUserKeyByIndex(firstUserId)
-          const secondUser = db.users.get(ws)
-
-          const gameId = Object.keys(db.games).length
-
-          if (secondUser) {
-            db.games[gameId] = {
-              id: gameId, 
-              users: [
-              { userId: firstUserId },
-              { userId: secondUser.index}
-            ]}
-          }
-
-          const firstUserResponse = toResponse<CreateGameResponseI>(
-            RESPONSE_TYPES.CREATE_GAME,
-            { idGame: gameId, idPlayer: 0 }
-          )
-          const secondUserResponse = toResponse<CreateGameResponseI>(
-            RESPONSE_TYPES.CREATE_GAME,
-            { idGame: gameId, idPlayer: 1 }
-          )
- 
-          if (firstUserWs) {
-            firstUserWs.send(firstUserResponse)
-          }
-          ws.send(secondUserResponse)
-
-          delete db.rooms[roomId];
-
-          const rooms = Object.values(db.rooms).map(r => ({ roomId: r.id, roomUsers: r.users }))
-          const updateRoomResponse = toResponse<UpdateRoomResponseI>(
-            RESPONSE_TYPES.UPDATE_ROOM, 
-            rooms
-          )
-          ws.send(updateRoomResponse)
-
+          addUserToRoom(ws, body, db);
+          updateRoom(ws, db);
           break;
         }
       }
